+++ packages/shared/storage/localStorage/helpers.ts
@@ -1,5 +1,5 @@
 import { LS_KEYS, getData, setData } from '.';
-import { Language } from '@ente/shared/i18n/locale';
+import { Language } from '@ente/shared/i18n';
 
 export const getToken = (): string => {
     const token = getData(LS_KEYS.USER)?.token;
